refactor: move rfc327 (commit formatting) check to allPRs
