+++ org/allPRs.ts
@@ -1,14 +1,13 @@
 import { danger, warn, fail, GitHubCommit, markdown } from "danger"
-import spellcheck from "danger-plugin-spellcheck"
-import yarn from "danger-plugin-yarn"
 
-import rfc327 from "./rfc_327"
+import yarn from "danger-plugin-yarn"
 
 // "Highlight package dependencies on Node projects"
 const rfc1 = async () => {
   await yarn()
 }
 
+import spellcheck from "danger-plugin-spellcheck"
 // "Keep our Markdown documents awesome",
 const rfc2 = async () => {
   await spellcheck({ settings: "artsy/peril-settings@spellcheck.json" })
@@ -218,6 +217,21 @@ export const deploySummary = async () => {
       .join("")
 
   return markdown(message)
+}
+
+// Nudge PR authors to use semantic commit formatting
+// https://github.com/artsy/README/issues/327
+export const rfc327 = () => {
+  const semanticFormat = /^(fix|feat|build|chore|ci|docs|style|refactor|perf|test)(?:\(.+\))?!?:.+$/
+
+  const pr = danger.github.pr
+  const repoName = pr.base.repo.name
+
+  if (["peril-settings", "volt"].includes(repoName) && !semanticFormat.test(pr.title)) {
+    return markdown(
+      "Hi there! :wave:\n\nWe're trialing semantic commit formatting which has not been detected in your PR title.\n\nRefer to [this RFC](https://github.com/artsy/README/issues/327#issuecomment-698842527) and [Conventional Commits](https://www.conventionalcommits.org) for PR/commit formatting guidelines."
+    )
+  }
 }
 
 // The default run
